ai-server/requirements.txt
tensorflow==2.15.0
numpy==1.26.0
opencv-python==4.8.1.78
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
mtcnn==0.1.1
pillow==10.1.0
scipy==1.11.3
pyjwt==2.8.0
